+++ src/Webiny/Component/TwitterOAuth/CHANGELOG.md
@@ -4,3 +4,9 @@ CHANGELOG
 1.0.0-RC1
 ---------
 * Initial version
+
+1.2.
+---------
+* removed the abraham/twitteroauth dependency
+* created a bridge for league/oauth1-client
+* refactored TwitterOAuthInterface